pyproject.toml
[project]
name = "notify-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "detect-secrets>=1.5.0",
    "fastapi[standard]>=0.135.1",
    "pyjwt>=2.12.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "sqlalchemy[asyncio]>=2.0.48",
]
